Think about behavior we need
Make a to-do list to remind us what we need to do, to keep us focused, and to tell us when we are finished.
When we start working on an item, we'll make it bold.
When we finish an item, we'll cross it off.
When we think of another test to write, we'll add it to the list.